requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
python-dotenv==1.0.1
bcrypt==4.1.2 